# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Sign Workflow API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents for the current user
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a new document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
                title:
                  type: string
                signing_order:
                  type: string
                  enum: [sequential, simultaneous]
              required:
                - document
      responses:
        "200":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDocumentResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with its recipients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /documents/{id}/fields:
    get:
      operationId: getDocumentFields
      tags: [documents]
      summary: Get signature field placements for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signature fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignatureFieldsResponse"
    put:
      operationId: saveDocumentFields
      tags: [documents]
      summary: Save signature field placements (replaces all)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFieldsRequest"
      responses:
        "200":
          description: Fields saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /documents/{id}/status:
    get:
      operationId: getDocumentStatus
      tags: [documents]
      summary: Get signing status for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStatusResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    SignatureField:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        recipientId:
          type: string
        page:
          type: integer
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        fieldType:
          type: string
          enum: [signature, initials, date, text]
        fieldValue:
          type: string
          nullable: true
      required:
        - id
        - documentId
        - recipientId
        - page
        - x
        - y
        - width
        - height
        - fieldType
    SignatureFieldInput:
      type: object
      properties:
        recipientId:
          type: string
        page:
          type: integer
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        fieldType:
          type: string
          enum: [signature, initials, date, text]
      required:
        - recipientId
        - page
        - x
        - y
        - width
        - height
        - fieldType
    SignatureFieldsResponse:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SignatureField"
      required:
        - fields
    SaveFieldsRequest:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SignatureFieldInput"
      required:
        - fields
    ReviewChecklistItem:
      type: object
      properties:
        label:
          type: string
        checked:
          type: boolean
      required:
        - label
        - checked
    Recipient:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        teamName:
          type: string
        email:
          type: string
        signOrder:
          type: integer
        status:
          type: string
          enum: [pending, viewed, signed]
        token:
          type: string
        signedAt:
          type: string
          nullable: true
        viewedAt:
          type: string
          nullable: true
        signerName:
          type: string
          nullable: true
        signatureData:
          type: string
          nullable: true
          description: Base64-encoded signature image (present when signed)
        createdAt:
          type: string
        requiresReview:
          type: boolean
        requiresSignature:
          type: boolean
        reviewStatus:
          type: string
          nullable: true
          enum: [pending, approved, changes_requested]
        reviewedAt:
          type: string
          nullable: true
        reviewNote:
          type: string
          nullable: true
        reviewChecklist:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ReviewChecklistItem"
      required:
        - id
        - documentId
        - teamName
        - email
        - signOrder
        - status
        - token
        - createdAt
        - requiresReview
        - requiresSignature
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        filename:
          type: string
        signingOrder:
          type: string
          enum: [sequential, simultaneous]
        status:
          type: string
          enum: [draft, in_review, sent, changes_requested, completed]
        uploadedBy:
          type: string
        uploaderName:
          type: string
        createdAt:
          type: string
        completedAt:
          type: string
          nullable: true
        totalRecipients:
          type: integer
        signedCount:
          type: integer
      required:
        - id
        - title
        - filename
        - signingOrder
        - status
        - uploadedBy
        - uploaderName
        - createdAt
        - totalRecipients
        - signedCount
    DocumentListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
      required:
        - documents
    UploadDocumentResponse:
      type: object
      properties:
        success:
          type: boolean
        documentId:
          type: string
      required:
        - success
        - documentId
    DocumentDetailResponse:
      type: object
      properties:
        document:
          $ref: "#/components/schemas/Document"
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/Recipient"
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SignatureField"
      required:
        - document
        - recipients
        - fields
...
    DocumentStatusResponse:
      type: object
      properties:
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/Recipient"
        status:
          type: string
      required:
        - recipients
        - status
...